html/semantics/embedded-content-0/media-elements/audio_property_set_value.html

<!doctype html>
<html>
 <head>
  <title>video.audio - set value</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#attr-media-audio">spec reference</a></p>
  <video id="video">
  </video>
  <div id="log"></div>
  <script>
test(function() {
  var v = document.getElementById("video");
  v.audio = "muted";
  assert_false(v.hasAttribute('audio'));
});
  </script>
 </body>
</html>
